% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_missing_demographic_data.R
\name{create_hash_id}
\alias{create_hash_id}
\title{Name: create_hash_id}
\usage{
create_hash_id(taf_df, id_colname = "hash_id")
}
\arguments{
\item{id_colname}{\itemize{
\item what you want to call the unique_id column - defaults to
hash_id
}}

\item{taf}{data frame (must include state_cd, msis_id, and bene_id columns)}
}
\value{
dataframe with unique hash_id for each beneficiary
}
\description{
Name: create_hash_id
}
