% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_missing_demographic_data.R
\name{fill_in_missing_demographic_data}
\alias{fill_in_missing_demographic_data}
\title{Name: fill_in_missing_demographic_data}
\usage{
fill_in_missing_demographic_data(demog_data, year_col, demog_col)
}
\arguments{
\item{demog_data}{\itemize{
\item dataframe or similar structure, containing the
taf_demog_elig_base
}}

\item{year_col}{\itemize{
\item name of the column that contains the year. TAF data does
not initially have a year column, as it comes in discreet files by year.
However, multiple, distinguishable years are necessary for this function. A
year column must be added.
}}

\item{demog_col}{\itemize{
\item name of the column with missing values you want to impute
(e.g. race_ethncty_cd, sex_cd, etc)
}}
}
\value{
dataframe with some missing observations in demog_col filled in in a
column called <demog_col>_imputed. Original values remain in the original column
}
\description{
Name: fill_in_missing_demographic_data
}
