% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_IL_claims.R
\name{fix_il_claims}
\alias{fix_il_claims}
\title{Collapses the IL claims data - which differs from that of other states -
to the most recent non-NA value of each field withing a 'claim family', per
resdac instrutions (which are available here: )}
\usage{
fix_il_claims(claims_dt)
}
\arguments{
\item{claims_dt}{A data table containing TAF claims with at least the fields
'state_cd', 'clm_type_cd', 'clm_num_adj', 'clm_num_orig', and 'adjdctn_dt'}
}
\description{
Collapses the IL claims data - which differs from that of other states -
to the most recent non-NA value of each field withing a 'claim family', per
resdac instrutions (which are available here: )
}
\examples{
il_dt_fixed <- fix_il_claims(il_dt)

}
\author{
Conor Hennessy
}
\keyword{~Claims}
\keyword{~Illinois}
\keyword{~TAF}
