% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apr_data_quality_measures.R
\name{get_apr_data_quality_measures}
\alias{get_apr_data_quality_measures}
\title{Name: get_apr_data_quality_measures}
\usage{
get_apr_data_quality_measures(
  taf_prvdr_base_df,
  taf_prvdr_taxonomy_df = NULL,
  taf_prvdr_location_df = NULL,
  year_col = NULL
)
}
\arguments{
\item{taf_prvdr_base_df}{\itemize{
\item dataframe, must have data from 'taf_prvdr_base' file
}}

\item{year_col}{\itemize{
\item a character or string with the name of the column with the year
if such a column exists (no such column is in the raw TAF data, but is necessary
if you are analyzing multiple years of data)
}}
}
\value{
A dataframe containing the same measures as the .csv files that can
be downloaded with the 'data (csv)' link at the URL above
}
\description{
Description: replication code for -
https://www.medicaid.gov/dq-atlas/landing/topics/single/map?topic=g6m94&tafVersionId=34
the following measures/topics are currently finished -
'Facility/Group/Individual Code'
}
