% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opioid_use_disorder_definitions.R
\name{get_icd10_codes_for_opioid_use_disorder}
\alias{get_icd10_codes_for_opioid_use_disorder}
\title{Name: get_icd10_codes_for_opioid_use_disorder}
\usage{
get_icd10_codes_for_opioid_use_disorder()
}
\value{
list of ICD10 codes as a dictionary with condition names as words and
ICD10 codes as defitions
}
\description{
Description: returns all ICD10 codes for opioid use disorder as a dictionary
ICD10 codes were chosen and named based on the information here:
https://www.icd10data.com/ICD10CM/Codes/F01-F99/F10-F19/F11-
}
