% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opioid_use_disorder_definitions.R
\name{get_icd9_codes_for_opioid_use_disorder}
\alias{get_icd9_codes_for_opioid_use_disorder}
\title{Name: get_icd9_codes_for_opioid_use_disorder}
\usage{
get_icd9_codes_for_opioid_use_disorder()
}
\value{
list of ICD9 codes as a dictionary with condition names as words and
ICD9 codes as defitions
}
\description{
Description: returns all ICD9 codes for opioid use disorder as a dictionary
ICD9 codes were chosen and named based on the information here:
https://www.ncbi.nlm.nih.gov/books/NBK557173/table/sb256.tab7/
}
