% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tennessee_leading_zeros.R
\name{remove_tennessee_icd_leading_zeros}
\alias{remove_tennessee_icd_leading_zeros}
\title{Name: remove_tennessee_zeros}
\usage{
remove_tennessee_icd_leading_zeros(tennessee_data, col = "all")
}
\value{
same data frame with leading zeros removed from specified column or
all dgns columns
}
\description{
Description: removes leading zeros from Tennessee ICD codes, given a
a dataframe with columns containing 'dgns'
}
\details{
defaults to removing leading zeros from all 'dgns' columns, but you can
specify a single column with the 'col' parameter
}
