% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michigan_county_fips_fix_17_18_19.R
\name{michigan_county_fips_fix_17_18_19}
\alias{michigan_county_fips_fix_17_18_19}
\title{Name: michigan_county_fips_fix_17_18_19}
\usage{
michigan_county_fips_fix_17_18_19(taf_demog_elig_base, year)
}
\value{
same data frame with county FIPS codes instead of SSA codes + 1 wherever
the bene_state_cd == 26 indicating Michigan
}
\description{
Description: the bene_cnty_cd variable in the 2017-2019 TAF data for Michagan
is incorrect, and appears to have the value of the county SSA code plus one,
rather than the FIPS code. This can be corrected using a publicly available
FIPS to SSA crosswalk. This function makes this correction, stores the original
bene_cnty_cd in a column called orig_bene_cnty_cd, and preserves the original
order of the data
}
